请求首页时，从redis查，如果小于等于4个活动，就全部取出传给客户端；如果大于四个活动，取前四个发给客户端

+++ backstage/src/main/java/com/example/demo/backstage/pojo/Activity.java
@@ -10,7 +10,7 @@ import java.math.BigInteger;
 
 @Data
 public class Activity {
-    private BigInteger id;
+    private Integer id;
     private String name;
     private String pic;
     private String link;
